pyproject.toml
[project]
name = "task-scheduler"
version = "1.0.0"
description = "Windows task scheduler for Python scripts and batch files"
requires-python = ">=3.11"
dependencies = [
    "APScheduler>=3.11.0",
    "bot-commander",
    "prompt_toolkit>=3.0.52",
    "tzdata>=2025.2",
    "tzlocal>=5.3.1",
    "wcwidth>=0.2.13",
    "pyftpsync>=4.0.0",
]

[project.optional-dependencies]
xmpp = ["xmpp-bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
xmpp-bot = { path = "../xmpp-bot", editable = true }
bot-commander = { path = "../bot-commander", editable = true }

[dependency-groups]
dev = [
    "pyinstaller>=6.0.0",
    "pytest>=9.0.2",
]
